feat: handle empty download directory
