+++ src/ui/components/download/DownloadFooter.tsx
@@ -1,0 +1,8 @@
+export const DownloadFooter = () => {
+  return (
+    <p>
+      This folder will be automatically deleted. Return to{' '}
+      <a href="https://2anki.net">2anki.net</a>
+    </p>
+  );
+};
