+++ src/ui/components/download/DownloadTitle.tsx
@@ -1,0 +1,7 @@
+interface DownloadTitleProps {
+  hasFiles: boolean;
+}
+
+export const DownloadTitle = ({ hasFiles }: DownloadTitleProps) => {
+  return hasFiles ? 'Your downloads are ready' : 'No downloads available';
+};
